updated package creation process to cope with npm ci
